// src/structures/product.type.ts
type ProductType = {
  id: String;
  appUUID: String;
  category: String;
  product: String;
  price: Number;
  currency: String;
  checkColor: Boolean;
  colors: String;
  description: String;
  images: String;
  published: Boolean;
  createdAt: String;
  updateAt: String;
};

export default ProductType;